refactor:move CanDetail.jsx to pages/; separate state from api
